+++ Modern JavaScript From The Beginning/Section 2: JavaScript Language Fundamentals/10. Numbers & The Math Object/app.js
@@ -1,0 +1,111 @@
+const firstNumber = 100;
+const secondNumber = 50;
+
+let value;
+
+console.log(value);
+console.log('---');
+
+// simple math with numbers
+
+value = firstNumber + secondNumber;
+
+console.log(value);
+console.log('---');
+
+value = firstNumber * secondNumber;
+
+console.log(value);
+console.log('---');
+
+value = firstNumber - secondNumber;
+
+console.log(value);
+console.log('---');
+
+value = firstNumber / secondNumber;
+
+console.log(value);
+console.log('---');
+
+value = firstNumber % secondNumber;
+
+console.log(value);
+console.log('---');
+
+// math object
+
+value = Math.PI;
+
+console.log(value);
+console.log('---');
+
+value = Math.E;
+
+console.log(value);
+console.log('---');
+
+value = Math.round(2.8);
+
+console.log(value);
+console.log('---');
+
+value = Math.round(2.4);
+
+console.log(value);
+console.log('---');
+
+value = Math.ceil(2.4);
+
+console.log(value);
+console.log('---');
+
+value = Math.floor(2.8);
+
+console.log(value);
+console.log('---');
+
+value = Math.sqrt(64);
+
+console.log(value);
+console.log('---');
+
+value = Math.abs(-3);
+
+console.log(value);
+console.log('---');
+
+value = Math.pow(8, 2);
+
+console.log(value);
+console.log('---');
+
+value = Math.min(2, 33, 4, 1, 55, 6, 3, -2);
+
+console.log(value);
+console.log('---');
+
+value = Math.max(2, 33, 4, 1, 55, 6, 3, -2);
+
+console.log(value);
+console.log('---');
+
+value = Math.random();
+
+console.log(value);
+console.log('---');
+
+value = Math.random() * 20;
+
+console.log(value);
+console.log('---');
+
+value = Math.random() * 20 + 1;
+
+console.log(value);
+console.log('---');
+
+value = Math.floor(Math.random() * 20 + 1);
+
+console.log(value);
+console.log('---');